Extract the environment handling into a separate abstract class.
